This People Need a Party
- 9 more NPCs were created.
-The "RageSlider" script was created. This script contains the logic for the rage slider's functionality.
-The "VirusSlider" script was created. This script contains the logic for the virus slider's functionality.

+++ Assets/_Project/Player/Projectile/Syringe.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class Syringe : MonoBehaviour
 {
-    public float healAmount; //50 es un buen valor.
+    public int healAmount; //50 es un buen valor.
 }
